ipoe,libnetlink: deal with client routes in vrf
When an interface from the [ipoe] configuration section belongs to a
VRF, add and delete the /32 routes to the clients in the interface VRF.

+++ accel-pppd/ctrl/ipoe/ipoe.h
@@ -52,7 +52,7 @@ struct ipoe_serv {
 #ifdef HAVE_VRF
 	int vrf_ifindex;
 	char vrf_name[IFNAMSIZ];
-	uint8_t table_id;
+	uint32_t table_id;
 #endif /* HAVE_VRF */
 	int vid;
 	int parent_vid;
